pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data_pipeline"
version = "0.1.0"
description = "Data pipeline for weather forecast ingestion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "requests",
    "dbt-core>=1.7.0",
    "dbt-clickhouse>=1.7.0",
    "boto3>=1.38.29",
    "apache-airflow>=3.0.2",
    "scaleway>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/data_pipeline"]
